updated CRUD methods in Task logic

+++ src/main/java/com/example/taskmanager/dao/impl/TaskDaoJdbcImpl.java
@@ -3,6 +3,7 @@ package com.example.taskmanager.dao.impl;
 import com.example.taskmanager.dao.TaskDao;
 import com.example.taskmanager.model.Task;
 import com.example.taskmanager.config.ConnectionUtil;
+import com.example.taskmanager.model.TaskStatus;
 
 import java.sql.*;
 import java.util.ArrayList;
@@ -14,7 +15,8 @@ public class TaskDaoJdbcImpl implements TaskDao {
     public Task create(Task task) {
         String sql = "INSERT INTO tasks (title, status, description, user_id) VALUES (?, ?, ?, ?)";
 
-        try (Connection connection = ConnectionUtil.getConnection(); PreparedStatement statement = connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
+        try (Connection connection = ConnectionUtil.getConnection();
+             PreparedStatement statement = connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
 
             statement.setString(1, task.getTitle());
             statement.setString(2, task.getStatus().name());
@@ -38,27 +40,31 @@ public class TaskDaoJdbcImpl implements TaskDao {
 
 
     @Override
-    public List<Task> getAll() {
-        String sql = "SELECT * FROM tasks";
+    public List<Task> getTasksByUserId(Long userId) {
+        String sql = "SELECT * FROM tasks WHERE user_id = ?";
         List<Task> tasks = new ArrayList<>();
         try (Connection connection = ConnectionUtil.getConnection();
-             PreparedStatement statement = connection.prepareStatement(sql);
-             ResultSet resultSet = statement.executeQuery()) {
-            while (resultSet.next()) {
-                tasks.add(getTask(resultSet));
+             PreparedStatement statement = connection.prepareStatement(sql)) {
+            statement.setLong(1, userId);
+
+            try (ResultSet resultSet = statement.executeQuery()) {
+                while (resultSet.next()) {
+                    tasks.add(getTask(resultSet));
+                }
+                return tasks;
             }
-            return tasks;
         } catch (SQLException e) {
             throw new RuntimeException("Can't get all tasks from DB", e);
         }
     }
 
     @Override
-    public Optional<Task> getById(Long id) {
-        String sql = "SELECT * FROM tasks WHERE id = ?";
+    public Optional<Task> getById(Long id, Long userId) {
+        String sql = "SELECT * FROM tasks WHERE id = ? AND user_id = ?";
         try (Connection connection = ConnectionUtil.getConnection(); PreparedStatement statement = connection.prepareStatement(sql)) {
 
             statement.setLong(1, id);
+            statement.setLong(2, userId);
 
             try (ResultSet resultSet = statement.executeQuery()) {
                 if (resultSet.next()) {
@@ -73,10 +79,12 @@ public class TaskDaoJdbcImpl implements TaskDao {
 
 
     @Override
-    public int delete(Long id) {
-        String sql = "DELETE FROM tasks WHERE id = ?";
+    public int delete(Long id, Long userId) {
+        String sql = "DELETE FROM tasks WHERE id = ? AND user_id = ?";
         try (Connection connection = ConnectionUtil.getConnection(); PreparedStatement statement = connection.prepareStatement(sql)) {
             statement.setLong(1, id);
+            statement.setLong(2, userId);
+
             int affectedRows = statement.executeUpdate();
             return affectedRows;
         } catch (SQLException e) {
@@ -86,13 +94,15 @@ public class TaskDaoJdbcImpl implements TaskDao {
 
     @Override
     public int update(Task task) {
-        String sql = "UPDATE tasks SET title = ?, completed = ? WHERE id = ?";
+        String sql = "UPDATE tasks SET title = ?, description = ?, status = ? WHERE id = ? AND user_id = ?";
 
         try (Connection connection = ConnectionUtil.getConnection(); PreparedStatement statement = connection.prepareStatement(sql)) {
 
             statement.setString(1, task.getTitle());
-            statement.setBoolean(2, task.isCompleted());
+            statement.setString(2, task.getStatus().name());
             statement.setLong(3, task.getId());
+            statement.setLong(4, task.getId());
+            statement.setLong(5, task.getUserId());
 
             int affectedRows = statement.executeUpdate();
             return affectedRows;
@@ -104,6 +114,8 @@ public class TaskDaoJdbcImpl implements TaskDao {
     private static Task getTask(ResultSet resultSet) throws SQLException {
         return new Task(resultSet.getLong("id"),
                 resultSet.getString("title"),
-                resultSet.getBoolean("completed"));
+                resultSet.getString("description"),
+                TaskStatus.fromString(resultSet.getString("status")),
+                resultSet.getLong("user_id"));
     }
 }
